Adding a bunch of exceptions
